backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
pgvector>=0.2.4

# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Content Extraction
trafilatura>=1.6.0
readability-lxml>=0.8.1
lxml>=5.1.0

# Keyword Extraction
yake>=0.4.8
kiwipiepy>=0.18.0  # Korean morphological analyzer

# ML - Embeddings
sentence-transformers>=2.2.2
torch>=2.2.0
numpy>=1.26.0

# ML - Clustering
hdbscan>=0.8.33
umap-learn>=0.5.5
scikit-learn>=1.4.0

# HTTP Client
httpx>=0.26.0

# Auth - JWT
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
